Move `isMarkedInternal` to symbols.ts
